requirements.txt
requests>=2.25.0
PyPDF2>=3.0.0
openai>=1.0.0
tqdm>=4.60.0
bs4
beautifulsoup4>=4.9.0
pymupdf>=1.23.0
boto3>=1.28.0
pillow>=10.0.0
